+++ src/components/moleculars/NewHeader/index.tsx
@@ -1,15 +1,12 @@
 import { useEffect, useState } from "react";
-
 import { useTranslation } from "react-i18next";
-
 import { logEvent } from "services/analytics";
-
 import { theme } from "@ribon.io/shared/styles";
 import Icon from "components/atomics/Icon";
 import RibonLogo from "./RibonLogo";
-import * as S from "./styles";
 import ConfigMenu from "../LayoutHeader/ConfigMenu";
 import TicketSection from "./TicketSection";
+import * as S from "./styles";
 
 function NewHeader(): JSX.Element {
   const { t } = useTranslation("translation", {
