+++ src/screens/donations/CausesScreen/index.tsx
@@ -1,32 +1,21 @@
 import { useCallback, useEffect, useRef, useState } from "react";
 import {
-  useStories,
   useFirstAccessToIntegration,
   useDonatedToday,
   useSubscriptions,
-  useReports,
 } from "@ribon.io/shared/hooks";
 import { ScrollView, Text, View, RefreshControl } from "react-native";
 import { useNavigation } from "hooks/useNavigation";
 import { useTranslation } from "react-i18next";
-import CardCenterImageButton from "components/moleculars/CardCenterImageButton";
 import GroupButtons from "components/moleculars/GroupButtons";
 import {
   INTEGRATION_AUTH_ID,
   RIBON_INTEGRATION_ID,
 } from "utils/constants/Application";
-import { NonProfit, Story } from "@ribon.io/shared/types";
-import StoriesSection from "screens/donations/CausesScreen/StoriesSection";
-import useFormattedImpactText from "hooks/useFormattedImpactText";
 import { logError } from "services/crashReport";
 import { useTicketsContext } from "contexts/ticketsContext";
-import ImpactDonationsVector from "screens/users/ImpactScreen/CommunityDonationsImpactCards/ImpactDonationsVector";
-import ZeroDonationsSection from "screens/users/ImpactScreen/ZeroDonationsSection";
 import { logEvent } from "services/analytics";
-import { Image as ExpoImage } from "expo-image";
-import InlineNotification from "components/moleculars/notifications/InlineNotification";
-import requestUserPermissionForNotifications from "lib/notifications";
-import { getLocalStorageItem, setLocalStorageItem } from "lib/localStorage";
+import { setLocalStorageItem } from "lib/localStorage";
 import { showToast } from "lib/Toast";
 import { useFocusEffect } from "@react-navigation/native";
 import * as SplashScreen from "expo-splash-screen";
@@ -50,17 +39,19 @@ import {
 import Placeholder from "./placeholder";
 import ContributionSection from "./ContributionSection";
 import DonationErrorModal from "./errorModalSection";
+import NonProfitsList from "./CausesSection/NonProfitsList";
 import ClubSection from "./ClubSection";
 import ReportsSection from "./ReportsSection";
+import NotificationPermissionPrompt from "./NotificationPermissionPrompt";
 import S from "./styles";
-
-const NOTIFICATION_CARD_VISIBLE_KEY = "NOTIFICATION_CARD_VISIBLE";
 
 export default function CausesScreen() {
   usePageView("P26_view");
+
   const { t } = useTranslation("translation", {
     keyPrefix: "donations.causesScreen",
   });
+
   const { nonProfitsWithPoolBalance: nonProfits, isLoading } =
     useNonProfitsContext();
   const { causesWithPoolBalance: causes } = useCausesContext();
@@ -76,21 +67,11 @@ export default function CausesScreen() {
   } = useFirstAccessToIntegration(currentIntegrationId);
   const { integration } = useIntegrationContext();
   const { ticketsCounter } = useTicketsContext();
-  const { reports, refetch: refetchReports } = useReports();
 
-  const [storiesVisible, setStoriesVisible] = useState(false);
-  const [stories, setStories] = useState<Story[]>([]);
-  const [currentNonProfit, setCurrentNonProfit] = useState<NonProfit>(
-    {} as NonProfit,
-  );
-  const [isNotificationCardVisible, setNotificationCardVisible] =
-    useState(false);
   const [refreshing, setRefreshing] = useState(false);
 
   const { navigateTo } = useNavigation();
   const scrollViewRef = useRef<any>(null);
-  const { fetchNonProfitStories } = useStories();
-  const { formattedImpactText } = useFormattedImpactText();
   const { hasTickets, refetchTickets } = useTicketsContext();
   const { currentUser, signedIn } = useCurrentUser();
   const { hasReceivedTicketToday, handleCanCollect, handleCollect } =
@@ -171,17 +152,6 @@ export default function CausesScreen() {
     ]),
   );
 
-  useEffect(() => {
-    const notificationCardVisible = async () => {
-      const value = await getLocalStorageItem(NOTIFICATION_CARD_VISIBLE_KEY);
-      return value === "true" || value === null;
-    };
-
-    notificationCardVisible().then((visible) => {
-      setNotificationCardVisible(visible);
-    });
-  }, []);
-
   const causesFilter = () => {
     const causesApi = causes.filter((cause) => cause.status === "active");
     return (
@@ -235,43 +205,6 @@ export default function CausesScreen() {
     sortNonProfits();
   }, [chosenCause]);
 
-  const handleNonProfitImagePress = async (nonProfit: NonProfit) => {
-    setCurrentNonProfit(nonProfit);
-    try {
-      const nonProfitStories = await fetchNonProfitStories(nonProfit.id);
-      ExpoImage.prefetch(nonProfitStories.map((story) => story.image));
-      if (nonProfitStories.length === 0) return;
-      setStories(nonProfitStories);
-      setStoriesVisible(true);
-      logEvent("storiesBtn_click", {
-        nonProfitId: nonProfit.id,
-        from: "NGOCard",
-      });
-    } catch (e) {
-      logError(e);
-    }
-  };
-
-  const nonProfitStylesFor = (index: number) => {
-    const isFirst = index === 0;
-    const isLast = index === nonProfitsFilter().length - 1;
-
-    return {
-      marginLeft: isFirst ? 16 : 4,
-      marginRight: isLast ? 16 : 4,
-      ...S.causesCardContainer,
-    };
-  };
-
-  const reportsStylesFor = (isMember: boolean) => ({
-    marginTop: 40,
-    marginBottom: isMember ? 48 : 8,
-  });
-
-  const navigateToClubScreen = () => {
-    navigateTo("ClubScreen");
-  };
-
   const shouldShowIntegrationBanner =
     !integration?.name?.toLowerCase()?.includes("ribon") &&
     integration &&
@@ -279,71 +212,17 @@ export default function CausesScreen() {
     hasTickets &&
     integration?.uniqueAddress !== INTEGRATION_AUTH_ID;
 
-  const handleHideNotificationClick = async () => {
-    const hideAlert = () => {
-      setLocalStorageItem(NOTIFICATION_CARD_VISIBLE_KEY, "false");
-      setNotificationCardVisible(false);
-    };
-
-    try {
-      const enabled = await requestUserPermissionForNotifications();
-      if (enabled) {
-        showToast({
-          type: "success",
-          message: t("enableNotification.successToastMessage"),
-          position: "bottom",
-        });
-        hideAlert();
-      }
-    } catch (e) {
-      logError(e);
-      showToast({
-        type: "error",
-        message: t("enableNotification.errorToastMessage"),
-        position: "bottom",
-      });
-      hideAlert();
-    }
-  };
-
-  const renderNotificationCard = () =>
-    isNotificationCardVisible && (
-      <View style={{ paddingBottom: 16 }}>
-        <InlineNotification
-          title={t("enableNotification.title")}
-          type="warning"
-          customIcon="notifications"
-          firstLink={t("enableNotification.link") || ""}
-          onFirstLinkClick={handleHideNotificationClick}
-        />
-      </View>
-    );
-
-  const handleButtonPress = (nonProfit: NonProfit) => {
-    logEvent("donateTicketBtn_start", {
-      nonProfitId: nonProfit.id,
-      from: "nonprofitCard",
-    });
-    if (signedIn) {
-      navigateTo("SelectTicketsScreen", {
-        nonProfit,
-        cause: nonProfit.cause,
-      });
-    } else {
-      navigateTo("DonationSignInScreen", { nonProfit });
-    }
-  };
-
   const { userIsMember } = useSubscriptions();
   const { isMember, refetch: refetchIsMember } = userIsMember();
 
   const onRefresh = async () => {
     setRefreshing(true);
     try {
-      refetchTickets();
-      await refetchIsMember();
-      await refetchFirstAccessToIntegration();
-      refetchReports();
+      await Promise.allSettled([
+        refetchTickets(),
+        refetchIsMember(),
+        refetchFirstAccessToIntegration(),
+      ]);
     } catch (e) {
       logError(e);
     } finally {
@@ -372,18 +251,10 @@ export default function CausesScreen() {
             },
           ]}
         >
-          {currentNonProfit && (
-            <StoriesSection
-              stories={stories}
-              nonProfit={currentNonProfit}
-              storiesVisible={storiesVisible}
-              setStoriesVisible={setStoriesVisible}
-            />
-          )}
           {shouldShowIntegrationBanner && (
             <IntegrationBanner integration={integration} />
           )}
-          {renderNotificationCard()}
+          <NotificationPermissionPrompt />
 
           {donatedToday && currentUser ? (
             <ContributionSection />
@@ -405,68 +276,8 @@ export default function CausesScreen() {
           </ScrollView>
         </View>
 
-        {sortNonProfits()?.length > 0 ? (
-          <ScrollView
-            style={S.causesContainer}
-            horizontal
-            showsHorizontalScrollIndicator={false}
-          >
-            {sortNonProfits()?.map((nonProfit, index) => {
-              const minNumberOfTickets =
-                nonProfit?.nonProfitImpacts?.[0]?.minimumNumberOfTickets ?? 0;
-              const hasEnoughTickets =
-                hasTickets && ticketsCounter >= minNumberOfTickets;
-              return (
-                <View style={nonProfitStylesFor(index)} key={nonProfit.id}>
-                  <CardCenterImageButton
-                    image={nonProfit.mainImage}
-                    infoTextTop={nonProfit.name}
-                    imageDescription={formattedImpactText(
-                      nonProfit,
-                      undefined,
-                      false,
-                      false,
-                      undefined,
-                      t("impactPrefix") || "",
-                    )}
-                    iconSubtitle={{
-                      icon: "confirmation_number",
-                      boldText: String(minNumberOfTickets),
-                      text: t("iconText"),
-                    }}
-                    buttonText={
-                      hasEnoughTickets ? t("buttonText") : t("notEnoughTickets")
-                    }
-                    onImagePress={() => {
-                      handleNonProfitImagePress(nonProfit);
-                    }}
-                    onClickButton={() => handleButtonPress(nonProfit)}
-                    buttonDisabled={!hasEnoughTickets}
-                    labelText={t("labelText") || ""}
-                  />
-                </View>
-              );
-            })}
-          </ScrollView>
-        ) : (
-          <View style={S.noCausesContainer}>
-            <ZeroDonationsSection
-              title={t("noCauses.title")}
-              onButtonPress={navigateToClubScreen}
-              description={t("noCauses.text")}
-              buttonText={t("noCauses.button")}
-              image={<ImpactDonationsVector />}
-            />
-          </View>
-        )}
-        {reports?.length ? (
-          <View>
-            <View style={S.divider} />
-            <View style={reportsStylesFor(isMember)}>
-              <ReportsSection data={reports} refetch={refetchReports} />
-            </View>
-          </View>
-        ) : null}
+        <NonProfitsList nonProfits={sortNonProfits()} />
+        <ReportsSection />
         <ClubSection isMember={isMember} refetch={refetchIsMember} />
       </ScrollView>
       <DonationErrorModal newState={params?.newState} />
